added rmStack
git-svn-id: file:///home/roberto/Git/svn@6309 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 61d7c585ea3f131332c6e6c6b5ba403d4ec3f85e

+++ trunk/xmipp/libraries/data/micrograph.cpp
@@ -464,7 +464,7 @@ void Micrograph::getLinearTransformatioVal8(double &a, double &b) const
 /* Produce all images ------------------------------------------------------ */
 void Micrograph::produce_all_images(int label, const FileName &fn_root,
                                     int starting_index, const FileName &fn_image, double ang, double tilt,
-                                    double psi)
+                                    double psi, bool rmStack)
 {
     ImageCollection SF(WRITE_APPEND);
     FileName fn_out;
@@ -513,8 +513,10 @@ void Micrograph::produce_all_images(int label, const FileName &fn_root,
     int nmax = ParticleNo();
     FileName fn_aux;
     fn_out          = fn_root;
-    if(exists(fn_out))
+    if(exists(fn_out) && rmStack)
+    {
         unlink(fn_out.c_str());
+    }
     int ii=0;
     for (int n = 0; n < nmax; n++)
         if (coords[n].valid && coords[n].label == label)
